refactor: change folder system and add loginregister form component
