+++ src/assets/styles/components/LoginRegistryForm/form.scss
@@ -1,0 +1,83 @@
+.form {
+  flex: 1 0 400px;
+  padding: 6rem 15px 15px;
+  background-color: #fff;
+
+  .form-inner {
+    display: block;
+    padding: 50px 25px;
+
+    .greeting {
+      font-size: clamp(0.5rem, 0.3vw + 1.3rem, 1.8rem);
+      margin-bottom: 20px;
+      width: max-content;
+    }
+
+    label {
+      display: block;
+      margin-bottom: 30px;
+      color: #aaa;
+      font-size: 1rem;
+      line-height: 1.5rem;
+      transition: 0.4s;
+      position: relative;
+
+      &:focus-within {
+        color: #c38f2c;
+
+        input {
+          background-color: #fff;
+          box-shadow: 0 0 6px rgba(0, 0, 0, 0.1);
+
+          &::placeholder {
+            color: #666;
+          }
+        }
+      }
+
+      input {
+        display: block;
+        width: 100%;
+        padding: 10px 15px;
+        border-radius: 10px;
+        border: #c0bfbf 1px solid;
+        position: relative;
+
+        color: #333;
+        font-size: 18px;
+
+        box-shadow: 0px 0px 0px rgba(0, 0, 0, 0);
+
+        transition: 0.4s;
+
+        &::placeholder {
+          color: #888;
+          transition: 0.4s;
+        }
+      }
+    }
+
+    button[type='submit'] {
+      display: block;
+      width: clamp(200px, 5vw, 65%);
+      padding: 10px 15px;
+      background-color: #ffc04c;
+      border-radius: 10px;
+      margin-inline: auto;
+
+      color: #fff;
+      font-size: 1.1rem;
+      font-weight: 500;
+      transition: background-color 0.5s ease;
+
+      &:hover:enabled {
+        background-color: #f5b133;
+      }
+
+      &:disabled {
+        opacity: 0.6;
+        cursor: auto;
+      }
+    }
+  }
+}
