+++ src/router/index.ts
@@ -1,27 +1,33 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import LoginPage from '@/components/LoginPage.vue';
-import ChatPage from '@/components/ChatPage/ChatPage.vue';
+import RegistryView from '@/views/RegistryView.vue';
+import LoginView from '@/views/LoginView.vue';
+import ChatView from '@/views/ChatView/ChatView.vue';
 import { useUserStore } from '@/stores/userStore';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
+      path: '/register',
+      name: 'Register',
+      component: RegistryView
+    },
+    {
       path: '/login',
       name: 'Login',
-      component: LoginPage
+      component: LoginView
     },
     {
       path: '/messages',
       name: 'Messages',
-      component: ChatPage
+      component: ChatView
     }
   ]
 });
 
 router.beforeEach((to) => {
   const store = useUserStore();
-  if (!store.user.isUserLoggedIn && to.name !== 'Login') {
+  if (!store.user.isUserLoggedIn && to.name !== 'Register' && to.name !== 'Login') {
     return { name: 'Login' };
   }
 });
