+++ src/views/RegistryView.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div class="registry"></div>
+</template>
+
+<style lang="scss"></style>
